pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "google-auth>=2.41.1",
    "gunicorn>=23.0.0",
    "pillow>=11.3.0",
    "pymongo>=4.15.3",
    "python-dotenv>=1.1.1",
    "replit-object-storage>=1.0.2",
    "requests>=2.32.5",
    "stripe>=13.0.1",
    "werkzeug>=3.1.3",
    "wtforms>=3.2.1",
]
